add custom TagWidget for use with the included TagField
This fixes a bug where the output would include commas (if editing an existing
tagged item). Because the TagField only accepts space-separated single-word
tags, the pre-populated comma-delimited value would fail to validate. Thus,
the form could not be submitted without manually removing all of the commas.
The field's output will now be a simple space-separated list.